+++ experimental/seq2seq/decode.py
@@ -1,17 +1,18 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-########################################################################
+#   Copyright (c) 2018 PaddlePaddle Authors. All Rights Reserved.
 #
-# Copyright (c) 2020 Baidu.com, Inc. All Rights Reserved
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
 #
-# File: test_decode.py
-# Author: chenxuyi(chenxuyi@baidu.com)
-# Date: 2020/05/08 20:55:15
+#     http://www.apache.org/licenses/LICENSE-2.0
 #
-########################################################################
-"""
-    Comment.
-"""
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+
 from __future__ import division
 from __future__ import absolute_import
 from __future__ import print_function
@@ -32,49 +33,14 @@ from ernie.modeling_ernie import ErnieModel, ErnieModelForPretraining
 from ernie.modeling_ernie import _build_linear, _build_ln, append_name
 from ernie.tokenizing_ernie import ErnieTokenizer
 
+from experimental.seq2seq.modeling_ernie_gen import ErnieModelForGeneration
+
 from propeller import log
 import propeller.paddle as propeller
 
 logging.getLogger().handlers[0]=log.handlers[0]
 logging.getLogger().setLevel(logging.DEBUG)
 log = logging.getLogger() 
-
-class ErnieModelForGeneration(ErnieModel):
-    def __init__(self, cfg, name=None):
-        cfg['return_additional_info'] = True
-        cfg['has_pooler'] = False
-        super(ErnieModelForGeneration, self).__init__(cfg, name=name)
-        initializer = F.initializer.TruncatedNormal(scale=cfg['initializer_range'])
-        d_model = cfg['hidden_size']
-        d_vocab = cfg['vocab_size']
-
-        self.mlm = _build_linear(d_model, d_model, append_name(name, 'mask_lm_trans_fc'), initializer, act=cfg['hidden_act'])
-        self.mlm_ln = _build_ln(d_model, name = append_name(name, 'mask_lm_trans'))
-        self.mlm_bias = L.create_parameter(
-                dtype='float32',
-                shape=[d_vocab], 
-                attr=F.ParamAttr(
-                    name=append_name(name, 'mask_lm_out_fc.b_0'), 
-                    initializer=F.initializer.Constant(value=0.0)
-                    ),
-                is_bias=True,
-            )
-
-    def forward(self, src_ids, *args, **kwargs):
-        _, encoded, info = ErnieModel.forward(self, src_ids, *args, **kwargs)
-        #log.debug('hidden_-1 %r'% L.reduce_mean(info['hiddens'][0]).numpy())
-        #log.debug('hidden_0 %r'% L.reduce_mean(info['hiddens'][1]).numpy())
-        encoded = self.mlm(encoded)
-        encoded = self.mlm_ln(encoded)
-        logits = L.matmul(encoded, self.word_emb.weight, transpose_y=True) + self.mlm_bias
-        return None, logits, info
-
-
-
-class ErnieGenTokenizer(ErnieTokenizer):
-    def __init__(self,  vocab, **kwargs):
-        super(ErnieGenTokenizer, self).__init__(vocab, mask_token=None, **kwargs)
-
 
 @np.vectorize
 def rev_lookup(i):
